-Improving error reporting for many units -Adding movt_s instruction -Updating sample Makefiles with new boost dependencies -Adding dragon sample scene -Adding flag in global memory so trax programs can determine if pack-split-axis is used
Former-commit-id: 321e792065c684fcf94a8b2789c893958e9958db

+++ llvm_trax/lib/trax_cpp.cpp
@@ -440,8 +440,8 @@ int main( int argc, char* argv[] ) {
   const boost::chrono::milliseconds milliSeconds = boost::chrono::duration_cast<boost::chrono::milliseconds>(timeElapsed);
 
   printf("\n----\nTotal runtime: %lu.%03lu seconds (%d:%d:%lu.%03lu)\n",
-         boost::chrono::duration_cast<boost::chrono::seconds>(timeElapsed2).count(), milliSeconds.count(),
-         hours.count(), minutes.count(), seconds.count(), milliSeconds.count());
+         (unsigned long)(boost::chrono::duration_cast<boost::chrono::seconds>(timeElapsed2).count()), (unsigned long)milliSeconds.count(),
+         hours.count(), minutes.count(), (unsigned long)seconds.count(), (unsigned long)milliSeconds.count());
   return 0;
 }
